move some tests to unit tests
