+++ contracts/service-registry/src/contract.rs
@@ -9,11 +9,8 @@ use crate::error::ContractError;
 use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 use crate::state::{AuthorizationState, BondingState, Config, Service, Worker, CONFIG, SERVICES};
 
-/*
-// version info for migration info
-const CONTRACT_NAME: &str = "crates.io:service-registry";
-const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-*/
+mod execute;
+mod query;
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
@@ -129,225 +126,6 @@ pub fn execute(
     .map_err(axelar_wasm_std::ContractError::from)
 }
 
-pub mod execute {
-    use crate::state;
-    use crate::state::{AuthorizationState, WORKERS, WORKERS_PER_CHAIN};
-    use connection_router_api::ChainName;
-
-    use super::*;
-
-    pub fn require_governance(deps: &DepsMut, info: MessageInfo) -> Result<(), ContractError> {
-        let config = CONFIG.load(deps.storage)?;
-        if config.governance != info.sender {
-            return Err(ContractError::Unauthorized);
-        }
-        Ok(())
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn register_service(
-        deps: DepsMut,
-        service_name: String,
-        service_contract: Addr,
-        min_num_workers: u16,
-        max_num_workers: Option<u16>,
-        min_worker_bond: Uint128,
-        bond_denom: String,
-        unbonding_period_days: u16,
-        description: String,
-    ) -> Result<Response, ContractError> {
-        let key = &service_name.clone();
-
-        SERVICES.update(
-            deps.storage,
-            key,
-            |service| -> Result<Service, ContractError> {
-                match service {
-                    None => Ok(Service {
-                        name: service_name,
-                        service_contract,
-                        min_num_workers,
-                        max_num_workers,
-                        min_worker_bond,
-                        bond_denom,
-                        unbonding_period_days,
-                        description,
-                    }),
-                    _ => Err(ContractError::ServiceAlreadyExists),
-                }
-            },
-        )?;
-
-        // Response with attributes? event?
-        Ok(Response::new())
-    }
-
-    pub fn update_worker_authorization_status(
-        deps: DepsMut,
-        workers: Vec<Addr>,
-        service_name: String,
-        auth_state: AuthorizationState,
-    ) -> Result<Response, ContractError> {
-        SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        for worker in workers {
-            WORKERS.update(
-                deps.storage,
-                (&service_name, &worker.clone()),
-                |sw| -> Result<Worker, ContractError> {
-                    match sw {
-                        Some(mut worker) => {
-                            worker.authorization_state = auth_state.clone();
-                            Ok(worker)
-                        }
-                        None => Ok(Worker {
-                            address: worker,
-                            bonding_state: BondingState::Unbonded,
-                            authorization_state: auth_state.clone(),
-                            service_name: service_name.clone(),
-                        }),
-                    }
-                },
-            )?;
-        }
-
-        Ok(Response::new())
-    }
-
-    pub fn bond_worker(
-        deps: DepsMut,
-        info: MessageInfo,
-        service_name: String,
-    ) -> Result<Response, ContractError> {
-        let service = SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        let bond = if !info.funds.is_empty() {
-            info.funds
-                .iter()
-                .find(|coin| coin.denom == service.bond_denom)
-                .ok_or(ContractError::WrongDenom)?
-                .amount
-        } else {
-            Uint128::zero() // sender can rebond currently unbonding funds by just sending no new funds
-        };
-
-        WORKERS.update(
-            deps.storage,
-            (&service_name.clone(), &info.sender.clone()),
-            |sw| -> Result<Worker, ContractError> {
-                match sw {
-                    Some(worker) => Ok(worker.add_bond(bond)?),
-                    None => Ok(Worker {
-                        address: info.sender,
-                        bonding_state: BondingState::Bonded { amount: bond },
-                        authorization_state: AuthorizationState::NotAuthorized,
-                        service_name,
-                    }),
-                }
-            },
-        )?;
-
-        Ok(Response::new())
-    }
-
-    pub fn register_chains_support(
-        deps: DepsMut,
-        info: MessageInfo,
-        service_name: String,
-        chains: Vec<ChainName>,
-    ) -> Result<Response, ContractError> {
-        SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        WORKERS
-            .may_load(deps.storage, (&service_name, &info.sender))?
-            .ok_or(ContractError::WorkerNotFound)?;
-
-        let _res =
-            state::register_chains_support(deps.storage, service_name.clone(), chains, info.sender);
-
-        Ok(Response::new())
-    }
-
-    pub fn deregister_chains_support(
-        deps: DepsMut,
-        info: MessageInfo,
-        service_name: String,
-        chains: Vec<ChainName>,
-    ) -> Result<Response, ContractError> {
-        SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        WORKERS
-            .may_load(deps.storage, (&service_name, &info.sender))?
-            .ok_or(ContractError::WorkerNotFound)?;
-
-        for chain in chains {
-            WORKERS_PER_CHAIN.remove(deps.storage, (&service_name, &chain, &info.sender));
-        }
-
-        Ok(Response::new())
-    }
-
-    pub fn unbond_worker(
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        service_name: String,
-    ) -> Result<Response, ContractError> {
-        SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        let worker = WORKERS
-            .may_load(deps.storage, (&service_name, &info.sender))?
-            .ok_or(ContractError::WorkerNotFound)?;
-
-        let can_unbond = true; // TODO: actually query the service to determine this value
-
-        let worker = worker.unbond(can_unbond, env.block.time)?;
-
-        WORKERS.save(deps.storage, (&service_name, &info.sender), &worker)?;
-
-        Ok(Response::new())
-    }
-
-    pub fn claim_stake(
-        deps: DepsMut,
-        env: Env,
-        info: MessageInfo,
-        service_name: String,
-    ) -> Result<Response, ContractError> {
-        let service = SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
-
-        let worker = WORKERS
-            .may_load(deps.storage, (&service_name, &info.sender))?
-            .ok_or(ContractError::WorkerNotFound)?;
-
-        let (worker, released_bond) =
-            worker.claim_stake(env.block.time, service.unbonding_period_days as u64)?;
-
-        WORKERS.save(deps.storage, (&service_name, &info.sender), &worker)?;
-
-        Ok(Response::new().add_message(BankMsg::Send {
-            to_address: info.sender.into(),
-            amount: [Coin {
-                denom: service.bond_denom,
-                amount: released_bond,
-            }]
-            .to_vec(), // TODO: isolate coins
-        }))
-    }
-}
-
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<Binary, ContractError> {
     match msg {
@@ -366,62 +144,1899 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<Binary, ContractErr
     }
 }
 
-pub mod query {
-    use connection_router_api::ChainName;
+#[cfg(test)]
+mod test {
+    use std::str::FromStr;
 
-    use crate::state::{WeightedWorker, WORKERS, WORKERS_PER_CHAIN, WORKER_WEIGHT};
+    use connection_router_api::ChainName;
+    use cosmwasm_std::{
+        coins, from_binary,
+        testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier, MockStorage},
+        BlockInfo, Empty, OwnedDeps,
+    };
+
+    use crate::state::{WeightedWorker, WORKER_WEIGHT};
 
     use super::*;
 
-    pub fn get_active_workers(
-        deps: Deps,
-        service_name: String,
-        chain_name: ChainName,
-    ) -> Result<Vec<WeightedWorker>, ContractError> {
-        let service = SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)?;
+    const GOVERNANCE_ADDRESS: &str = "governance";
+    const UNAUTHORIZED_ADDRESS: &str = "unauthorized";
+    const WORKER_ADDRESS: &str = "worker";
+    const AXL_DENOMINATION: &str = "uaxl";
 
-        let workers: Vec<_> = WORKERS_PER_CHAIN
-            .prefix((&service_name, &chain_name))
-            .range(deps.storage, None, None, Order::Ascending)
-            .map(|res| res.and_then(|(addr, _)| WORKERS.load(deps.storage, (&service_name, &addr))))
-            .collect::<Result<Vec<Worker>, _>>()?
-            .into_iter()
-            .filter(|worker| match worker.bonding_state {
-                BondingState::Bonded { amount } => amount >= service.min_worker_bond,
-                _ => false,
-            })
-            .filter(|worker| worker.authorization_state == AuthorizationState::Authorized)
-            .map(|worker| WeightedWorker {
-                worker_info: worker,
-                weight: WORKER_WEIGHT, // all workers have an identical const weight for now
-            })
-            .collect();
+    fn setup() -> OwnedDeps<MockStorage, MockApi, MockQuerier, Empty> {
+        let mut deps = mock_dependencies();
 
-        if workers.len() < service.min_num_workers.into() {
-            Err(ContractError::NotEnoughWorkers)
-        } else {
-            Ok(workers)
+        instantiate(
+            deps.as_mut(),
+            mock_env(),
+            mock_info("instantiator", &[]),
+            InstantiateMsg {
+                governance_account: GOVERNANCE_ADDRESS.to_string(),
+            },
+        );
+
+        deps
+    }
+
+    pub fn assert_contract_err_strings_equal(
+        actual: impl Into<axelar_wasm_std::ContractError>,
+        expected: impl Into<axelar_wasm_std::ContractError>,
+    ) {
+        assert_eq!(actual.into().to_string(), expected.into().to_string());
+    }
+
+    #[test]
+    fn register_service() {
+        let mut deps = setup();
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: "validators".into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond: Uint128::zero(),
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let err = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(UNAUTHORIZED_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: "validators".into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond: Uint128::zero(),
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        )
+        .unwrap_err();
+        assert_contract_err_strings_equal(err, ContractError::Unauthorized);
+    }
+
+    #[test]
+    fn authorize_worker() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond: Uint128::zero(),
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![Addr::unchecked("worker").into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let err = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(UNAUTHORIZED_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![Addr::unchecked("worker").into()],
+                service_name: service_name.into(),
+            },
+        )
+        .unwrap_err();
+        assert_contract_err_strings_equal(err, ContractError::Unauthorized);
+    }
+
+    #[test]
+    fn bond_worker() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+    }
+
+    #[test]
+    fn register_chain_support() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            workers,
+            vec![WeightedWorker {
+                worker_info: Worker {
+                    address: Addr::unchecked(WORKER_ADDRESS),
+                    bonding_state: BondingState::Bonded {
+                        amount: min_worker_bond
+                    },
+                    authorization_state: AuthorizationState::Authorized,
+                    service_name: service_name.into()
+                },
+                weight: WORKER_WEIGHT
+            }]
+        );
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name: ChainName::from_str("random chain").unwrap(),
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![]);
+    }
+
+    /// If a bonded and authorized worker deregisters support for a chain they previously registered support for,
+    /// that worker should no longer be part of the active worker set for that chain
+    #[test]
+    fn register_and_deregister_support_for_single_chain() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        // Deregister chain support
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![]);
+    }
+
+    /// Same setting and goal as register_and_deregister_support_for_single_chain() but for multiple chains.
+    #[test]
+    fn register_and_deregister_support_for_multiple_chains() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chains = vec![
+            ChainName::from_str("ethereum").unwrap(),
+            ChainName::from_str("binance").unwrap(),
+            ChainName::from_str("avalanche").unwrap(),
+        ];
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: chains.clone(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: chains.clone(),
+            },
+        );
+        assert!(res.is_ok());
+
+        for chain in chains {
+            let workers: Vec<WeightedWorker> = from_binary(
+                &query(
+                    deps.as_ref(),
+                    mock_env(),
+                    QueryMsg::GetActiveWorkers {
+                        service_name: service_name.into(),
+                        chain_name: chain,
+                    },
+                )
+                .unwrap(),
+            )
+            .unwrap();
+            assert_eq!(workers, vec![]);
         }
     }
 
-    pub fn get_worker(
-        deps: Deps,
-        service_name: String,
-        worker: String,
-    ) -> Result<Worker, ContractError> {
-        WORKERS
-            .may_load(
-                deps.storage,
-                (&service_name, &deps.api.addr_validate(&worker)?),
-            )?
-            .ok_or(ContractError::WorkerNotFound)
+    /// If a bonded and authorized worker deregisters support for the first chain among multiple chains,
+    /// they should remain part of the active worker set for all chains except the first one.
+    #[test]
+    fn register_for_multiple_chains_deregister_for_first_one() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chains = vec![
+            ChainName::from_str("ethereum").unwrap(),
+            ChainName::from_str("binance").unwrap(),
+            ChainName::from_str("avalanche").unwrap(),
+        ];
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: chains.clone(),
+            },
+        );
+        assert!(res.is_ok());
+
+        // Deregister only the first chain
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chains[0].clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        // Verify that worker is not associated with the deregistered chain
+        let deregistered_chain = chains[0].clone();
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name: deregistered_chain,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![]);
+
+        // Verify that worker is still associated with other chains
+        for chain in chains.iter().skip(1) {
+            let workers: Vec<WeightedWorker> = from_binary(
+                &query(
+                    deps.as_ref(),
+                    mock_env(),
+                    QueryMsg::GetActiveWorkers {
+                        service_name: service_name.into(),
+                        chain_name: chain.clone(),
+                    },
+                )
+                .unwrap(),
+            )
+            .unwrap();
+            assert_eq!(
+                workers,
+                vec![WeightedWorker {
+                    worker_info: Worker {
+                        address: Addr::unchecked(WORKER_ADDRESS),
+                        bonding_state: BondingState::Bonded {
+                            amount: min_worker_bond
+                        },
+                        authorization_state: AuthorizationState::Authorized,
+                        service_name: service_name.into()
+                    },
+                    weight: WORKER_WEIGHT
+                }]
+            );
+        }
     }
 
-    pub fn get_service(deps: Deps, service_name: String) -> Result<Service, ContractError> {
-        SERVICES
-            .may_load(deps.storage, &service_name)?
-            .ok_or(ContractError::ServiceNotFound)
+    /// If a bonded and authorized worker registers support for one chain and later deregisters support for another chain,
+    /// the active worker set for the original chain should remain unaffected by the deregistration.
+    #[test]
+    fn register_support_for_a_chain_deregister_support_for_another_chain() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let second_chain_name = ChainName::from_str("avalanche").unwrap();
+        // Deregister support for another chain
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![second_chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            workers,
+            vec![WeightedWorker {
+                worker_info: Worker {
+                    address: Addr::unchecked(WORKER_ADDRESS),
+                    bonding_state: BondingState::Bonded {
+                        amount: min_worker_bond
+                    },
+                    authorization_state: AuthorizationState::Authorized,
+                    service_name: service_name.into()
+                },
+                weight: WORKER_WEIGHT
+            }]
+        );
     }
+
+    /// If a bonded and authorized worker registers, deregisters, and again registers their support for a single chain,
+    /// the active worker set of that chain should include the worker.
+    #[test]
+    fn register_deregister_register_support_for_single_chain() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        // Second support declaration
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            workers,
+            vec![WeightedWorker {
+                worker_info: Worker {
+                    address: Addr::unchecked(WORKER_ADDRESS),
+                    bonding_state: BondingState::Bonded {
+                        amount: min_worker_bond
+                    },
+                    authorization_state: AuthorizationState::Authorized,
+                    service_name: service_name.into()
+                },
+                weight: WORKER_WEIGHT
+            }]
+        );
+    }
+
+    /// If a bonded and authorized worker deregisters their support for a chain they have not previously registered
+    /// support for, the call should be ignored and the active worker set of the chain should be intact.
+    #[test]
+    fn deregister_previously_unsupported_single_chain() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![])
+    }
+
+    /// If an unbonded but authorized worker deregisters support for a chain they previously registered support for,
+    /// that worker should not be part of the active worker set for that chain.
+    #[test]
+    fn register_and_deregister_support_for_single_chain_unbonded() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![]);
+    }
+
+    /// If a worker that is not part of a service deregisters support for a chain from that specific service,
+    /// process should return a contract error of type WorkerNotFound.
+    #[test]
+    fn deregister_from_unregistered_worker_single_chain() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let err = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        )
+        .unwrap_err();
+
+        assert_contract_err_strings_equal(err, ContractError::WorkerNotFound);
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![]);
+    }
+
+    /// If a worker deregisters support for a chain of an unregistered service,
+    /// process should return a contract error of type ServiceNotFound.
+    #[test]
+    fn deregister_single_chain_for_nonexistent_service() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let err = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::DeregisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        )
+        .unwrap_err();
+
+        assert_contract_err_strings_equal(err, ContractError::ServiceNotFound);
+    }
+
+    #[test]
+    fn unbond_worker() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::UnbondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![])
+    }
+
+    #[test]
+    fn bond_wrong_denom() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let err = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &coins(min_worker_bond.u128(), "funnydenom")),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        )
+        .unwrap_err();
+
+        assert_contract_err_strings_equal(err, ContractError::WrongDenom);
+    }
+
+    #[test]
+    fn bond_but_not_authorized() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![])
+    }
+
+    #[test]
+    fn bond_but_not_enough() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128() / 2, AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(workers, vec![])
+    }
+
+    #[test]
+    fn bond_before_authorize() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            workers,
+            vec![WeightedWorker {
+                worker_info: Worker {
+                    address: Addr::unchecked(WORKER_ADDRESS),
+                    bonding_state: BondingState::Bonded {
+                        amount: min_worker_bond
+                    },
+                    authorization_state: AuthorizationState::Authorized,
+                    service_name: service_name.into()
+                },
+                weight: WORKER_WEIGHT
+            }]
+        );
+    }
+
+    #[test]
+    fn unbond_then_rebond() {
+        let mut deps = setup();
+
+        let service_name = "validators";
+        let min_worker_bond = Uint128::new(100);
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::RegisterService {
+                service_name: service_name.into(),
+                service_contract: Addr::unchecked("nowhere"),
+                min_num_workers: 0,
+                max_num_workers: Some(100),
+                min_worker_bond,
+                bond_denom: AXL_DENOMINATION.into(),
+                unbonding_period_days: 10,
+                description: "Some service".into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(GOVERNANCE_ADDRESS, &[]),
+            ExecuteMsg::AuthorizeWorkers {
+                workers: vec![WORKER_ADDRESS.into()],
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(
+                WORKER_ADDRESS,
+                &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+            ),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let chain_name = ChainName::from_str("ethereum").unwrap();
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::RegisterChainSupport {
+                service_name: service_name.into(),
+                chains: vec![chain_name.clone()],
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::UnbondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            mock_info(WORKER_ADDRESS, &[]),
+            ExecuteMsg::BondWorker {
+                service_name: service_name.into(),
+            },
+        );
+        assert!(res.is_ok());
+
+        let workers: Vec<WeightedWorker> = from_binary(
+            &query(
+                deps.as_ref(),
+                mock_env(),
+                QueryMsg::GetActiveWorkers {
+                    service_name: service_name.into(),
+                    chain_name,
+                },
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            workers,
+            vec![WeightedWorker {
+                worker_info: Worker {
+                    address: Addr::unchecked(WORKER_ADDRESS),
+                    bonding_state: BondingState::Bonded {
+                        amount: min_worker_bond
+                    },
+                    authorization_state: AuthorizationState::Authorized,
+                    service_name: service_name.into()
+                },
+                weight: WORKER_WEIGHT
+            }]
+        );
+    }
+
+    // #[test]
+    // fn unbonding_period() {
+    //     let mut deps = setup();
+
+    //     let min_worker_bond = Uint128::new(100);
+    //     let initial_bal = min_worker_bond.u128() * 2;
+
+    //     let service_name = "validators";
+    //     let unbonding_period_days = 1;
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(GOVERNANCE_ADDRESS, &[]),
+    //         ExecuteMsg::RegisterService {
+    //             service_name: service_name.into(),
+    //             service_contract: Addr::unchecked("nowhere"),
+    //             min_num_workers: 0,
+    //             max_num_workers: Some(100),
+    //             min_worker_bond,
+    //             bond_denom: AXL_DENOMINATION.into(),
+    //             unbonding_period_days,
+    //             description: "Some service".into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(GOVERNANCE_ADDRESS, &[]),
+    //         ExecuteMsg::AuthorizeWorkers {
+    //             workers: vec![WORKER_ADDRESS.into()],
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(
+    //             WORKER_ADDRESS,
+    //             &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+    //         ),
+    //         ExecuteMsg::BondWorker {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     let chain_name = ChainName::from_str("ethereum").unwrap();
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(WORKER_ADDRESS, &[]),
+    //         ExecuteMsg::RegisterChainSupport {
+    //             service_name: service_name.into(),
+    //             chains: vec![chain_name],
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(WORKER_ADDRESS, &[]),
+    //         ExecuteMsg::UnbondWorker {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+    //     assert_eq!(
+    //         app.wrap()
+    //             .query_balance(WORKER_ADDRESS, AXL_DENOMINATION)
+    //             .unwrap()
+    //             .amount
+    //             .u128(),
+    //         initial_bal - min_worker_bond.u128()
+    //     );
+
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(WORKER_ADDRESS, &[]),
+    //         ExecuteMsg::ClaimStake {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_err());
+    //     assert_eq!(
+    //         res.unwrap_err().to_string(),
+    //         axelar_wasm_std::ContractError::from(ContractError::InvalidBondingState(
+    //             BondingState::Unbonding {
+    //                 unbonded_at: app.block_info().time,
+    //                 amount: min_worker_bond,
+    //             }
+    //         ))
+    //         .to_string()
+    //     );
+    //     assert_eq!(
+    //         app.wrap()
+    //             .query_balance(WORKER_ADDRESS, AXL_DENOMINATION)
+    //             .unwrap()
+    //             .amount
+    //             .u128(),
+    //         initial_bal - min_worker_bond.u128()
+    //     );
+
+    //     let block = app.block_info();
+    //     app.set_block(BlockInfo {
+    //         height: block.height + 1,
+    //         time: app
+    //             .block_info()
+    //             .time
+    //             .plus_days(unbonding_period_days.into()),
+    //         ..block
+    //     });
+
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(WORKER_ADDRESS, &[]),
+    //         ExecuteMsg::ClaimStake {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     assert_eq!(
+    //         app.wrap()
+    //             .query_balance(worker, AXL_DENOMINATION)
+    //             .unwrap()
+    //             .amount
+    //             .u128(),
+    //         initial_bal
+    //     );
+    // }
+
+    // #[test]
+    // fn get_active_workers_should_not_return_less_than_min() {
+    //     let workers = vec![Addr::unchecked("worker1"), Addr::unchecked("worker2")];
+    //     let min_num_workers = workers.len() as u16;
+
+    //     let mut app = App::new(|router, _, storage| {
+    //         for worker in &workers {
+    //             router
+    //                 .bank
+    //                 .init_balance(storage, worker, coins(100000, AXL_DENOMINATION))
+    //                 .unwrap()
+    //         }
+    //     });
+
+    //     let governance = Addr::unchecked("gov");
+    //     let service_registry =
+    //         test_utils::ServiceRegistryContract::instantiate_contract(&mut app, governance.clone());
+
+    //     let service_name = "validators";
+    //     let min_worker_bond = Uint128::new(100);
+    //     let _ = service_registry
+    //         .execute(
+    //             &mut app,
+    //             governance.clone(),
+    //             ExecuteMsg::RegisterService {
+    //                 service_name: service_name.into(),
+    //                 service_contract: Addr::unchecked("nowhere"),
+    //                 min_num_workers,
+    //                 max_num_workers: Some(100),
+    //                 min_worker_bond,
+    //                 bond_denom: AXL_DENOMINATION.into(),
+    //                 unbonding_period_days: 10,
+    //                 description: "Some service".into(),
+    //             },
+    //         )
+    //         .unwrap();
+
+    //     let _ = service_registry
+    //         .execute(
+    //             &mut app,
+    //             governance,
+    //             ExecuteMsg::AuthorizeWorkers {
+    //                 workers: workers.iter().map(|w| w.into()).collect(),
+    //                 service_name: service_name.into(),
+    //             },
+    //         )
+    //         .unwrap();
+
+    //     let chain_name = ChainName::from_str("ethereum").unwrap();
+
+    //     for worker in &workers {
+    //         // should return err until all workers are registered
+    //         let res: StdResult<Vec<WeightedWorker>> = service_registry.query(
+    //             &app,
+    //             QueryMsg::GetActiveWorkers {
+    //                 service_name: service_name.into(),
+    //                 chain_name: chain_name.clone(),
+    //             },
+    //         );
+    //         assert!(res.is_err());
+
+    //         let _ = service_registry
+    //             .execute_with_funds(
+    //                 &mut app,
+    //                 worker.clone(),
+    //                 ExecuteMsg::BondWorker {
+    //                     service_name: service_name.into(),
+    //                 },
+    //                 &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+    //             )
+    //             .unwrap();
+
+    //         let _ = service_registry
+    //             .execute(
+    //                 &mut app,
+    //                 worker.clone(),
+    //                 ExecuteMsg::RegisterChainSupport {
+    //                     service_name: service_name.into(),
+    //                     chains: vec![chain_name.clone()],
+    //                 },
+    //             )
+    //             .unwrap();
+    //     }
+
+    //     // all workers registered, should not return err now
+    //     let res: StdResult<Vec<WeightedWorker>> = service_registry.query(
+    //         &app,
+    //         QueryMsg::GetActiveWorkers {
+    //             service_name: service_name.into(),
+    //             chain_name: chain_name.clone(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     // remove one, should return err again
+    //     let _ = service_registry
+    //         .execute(
+    //             &mut app,
+    //             workers[0].clone(),
+    //             ExecuteMsg::DeregisterChainSupport {
+    //                 service_name: service_name.into(),
+    //                 chains: vec![chain_name.clone()],
+    //             },
+    //         )
+    //         .unwrap();
+    //     let res: StdResult<Vec<WeightedWorker>> = service_registry.query(
+    //         &app,
+    //         QueryMsg::GetActiveWorkers {
+    //             service_name: service_name.into(),
+    //             chain_name: chain_name.clone(),
+    //         },
+    //     );
+    //     assert!(res.is_err());
+    // }
+
+    // #[test]
+    // fn jail_worker() {
+    //     let faucet = Addr::unchecked("faucet");
+    //     // init app with faucet balance
+    //     let mut app = App::new(|router, _, storage| {
+    //         router
+    //             .bank
+    //             .init_balance(storage, &faucet, coins(100000, AXL_DENOMINATION))
+    //             .unwrap()
+    //     });
+
+    //     // init service registry contract
+    //     let governance = Addr::unchecked("gov");
+    //     let service_registry =
+    //         test_utils::ServiceRegistryContract::instantiate_contract(&mut app, governance.clone());
+
+    //     // register a service
+    //     let service_name = "validators";
+    //     let min_worker_bond = Uint128::new(100);
+    //     let unbonding_period_days = 10;
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(GOVERNANCE_ADDRESS, &[]),
+    //         ExecuteMsg::RegisterService {
+    //             service_name: service_name.into(),
+    //             service_contract: Addr::unchecked("service contract"),
+    //             min_num_workers: 0,
+    //             max_num_workers: Some(100),
+    //             min_worker_bond,
+    //             bond_denom: AXL_DENOMINATION.into(),
+    //             unbonding_period_days,
+    //             description: "Some service".into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     // given a bonded worker
+    //     let worker1 = Addr::unchecked("worker-1");
+    //     // fund worker
+    //     let msg: cosmwasm_std::CosmosMsg = BankMsg::Send {
+    //         to_address: worker1.clone().into(),
+    //         amount: coins(min_worker_bond.u128(), AXL_DENOMINATION),
+    //     }
+    //     .into();
+    //     app.execute(faucet.clone(), msg.clone()).unwrap();
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(
+    //             worker1.clone(),
+    //             &coins(
+    //                 &coins(min_worker_bond.u128(), AXL_DENOMINATION).u128(),
+    //                 AXL_DENOMINATION,
+    //             ),
+    //         ),
+    //         ExecuteMsg::BondWorker {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     // when worker is jailed
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(GOVERNANCE_ADDRESS, &[]),
+    //         ExecuteMsg::JailWorkers {
+    //             workers: vec![worker1.clone().into()],
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     // worker cannot unbond
+    //     let err = service_registry
+    //         .execute(
+    //             &mut app,
+    //             worker1.clone(),
+    //             ExecuteMsg::UnbondWorker {
+    //                 service_name: service_name.into(),
+    //             },
+    //         )
+    //         .unwrap_err();
+    //     assert_contract_err_strings_equal(err, ContractError::WorkerJailed);
+
+    //     // given a worker passed unbonding period
+    //     let worker2 = Addr::unchecked("worker-2");
+    //     // fund worker
+    //     let msg: cosmwasm_std::CosmosMsg = BankMsg::Send {
+    //         to_address: worker2.clone().into(),
+    //         amount: coins(min_worker_bond.u128(), AXL_DENOMINATION),
+    //     }
+    //     .into();
+    //     app.execute(faucet.clone(), msg.clone()).unwrap();
+    //     // bond worker
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(
+    //             worker2.clone(),
+    //             &coins(min_worker_bond.u128(), AXL_DENOMINATION),
+    //         ),
+    //         ExecuteMsg::BondWorker {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+    //     // unbond worker
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(worker2.clone(), &[]),
+    //         ExecuteMsg::UnbondWorker {
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+    //     let worker: Worker = query(
+    //         deps.as_ref(),
+    //         mock_env(),
+    //         QueryMsg::GetWorker {
+    //             service_name: service_name.into(),
+    //             worker: worker2.to_string(),
+    //         },
+    //     )
+    //     .unwrap();
+
+    //     let block = app.block_info();
+    //     assert_eq!(
+    //         worker.bonding_state,
+    //         BondingState::Unbonding {
+    //             amount: min_worker_bond,
+    //             unbonded_at: block.time,
+    //         }
+    //     );
+
+    //     // when worker is jailed
+    //     let res = execute(
+    //         deps.as_mut(),
+    //         mock_env(),
+    //         mock_info(GOVERNANCE_ADDRESS, &[]),
+    //         ExecuteMsg::JailWorkers {
+    //             workers: vec![worker2.clone().into()],
+    //             service_name: service_name.into(),
+    //         },
+    //     );
+    //     assert!(res.is_ok());
+
+    //     // and unbonding period has passed
+    //     app.set_block(BlockInfo {
+    //         height: block.height + 1,
+    //         time: block.time.plus_days((unbonding_period_days + 1).into()),
+    //         ..block
+    //     });
+
+    //     // worker cannot claim stake
+    //     let err = service_registry
+    //         .execute(
+    //             &mut app,
+    //             worker2.clone(),
+    //             ExecuteMsg::ClaimStake {
+    //                 service_name: service_name.into(),
+    //             },
+    //         )
+    //         .unwrap_err();
+    //     assert_contract_err_strings_equal(err, ContractError::WorkerJailed);
+    // }
 }
